Related work items: #19047, #19548, #19781, #19799, #19807, #20111, #20113, #20601, #20604, #20608, #20656, #20666, #20669

+++ dotnet/corecomponents/net.atos.daf.ct2.fms/entity/VehicleStatusResponse.cs
@@ -28,7 +28,7 @@ namespace net.atos.daf.ct2.fms.entity
     public class VehiclePositionForStatus
     {
         public int Altitude { get; set; }
-        public double Heading { get; set; }
+        public int Heading { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
         public long GPSTimestamp { get; set; }
